# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/session/{sessionId}/message/read": {
      "put": {
        "tags": [
          "消息管理"
        ],
        "summary": "标记已读",
        "operationId": "markRead",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMapStringObject"
                }
              }
            }
          }
        }
      }
    },
    "/goods/update": {
      "put": {
        "tags": [
          "商品管理"
        ],
        "summary": "更新商品",
        "operationId": "update",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGoodsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseGoodsResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/goods/add": {
      "post": {
        "tags": [
          "商品管理"
        ],
        "summary": "添加商品",
        "operationId": "add",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddGoodsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseGoodsResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/session/{sessionId}": {
      "get": {
        "tags": [
          "会话管理"
        ],
        "summary": "查询会话详情",
        "operationId": "detail",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseSessionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/session/{sessionId}/message/unread-count": {
      "get": {
        "tags": [
          "消息管理"
        ],
        "summary": "未读消息数",
        "operationId": "unreadCount",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMapStringObject"
                }
              }
            }
          }
        }
      }
    },
    "/session/{sessionId}/message/list": {
      "get": {
        "tags": [
          "消息管理"
        ],
        "summary": "查询消息列表",
        "operationId": "list",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "pageNum",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageMessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/session/user/list": {
      "get": {
        "tags": [
          "会话管理"
        ],
        "summary": "用户查询会话列表",
        "operationId": "listByUser",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageNum",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageSessionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/session/ct/list": {
      "get": {
        "tags": [
          "会话管理"
        ],
        "summary": "商户查询会话列表",
        "operationId": "listByTenant",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageNum",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageSessionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/goods/mine": {
      "get": {
        "tags": [
          "商品管理"
        ],
        "summary": "我的商品列表",
        "operationId": "mine",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageNum",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageGoodsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/goods/detail": {
      "get": {
        "tags": [
          "商品管理"
        ],
        "summary": "商品详情",
        "operationId": "detail_1",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseGoodsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/goods/all": {
      "get": {
        "tags": [
          "商品管理"
        ],
        "summary": "全部商品列表",
        "operationId": "all",
        "parameters": [
          {
            "name": "pageNum",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageGoodsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/goods/delete": {
      "delete": {
        "tags": [
          "商品管理"
        ],
        "summary": "删除商品",
        "operationId": "delete",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiResponseMapStringObject": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码：0=成功，1001=账号已存在，1002=账号不存在，1003=密码错误，1004=参数错误",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "success"
          },
          "data": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "description": "响应数据"
            },
            "description": "响应数据"
          }
        },
        "description": "通用响应"
      },
      "UpdateGoodsRequest": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "商品ID",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "商品名称",
            "example": "更新后的商品名称"
          }
        },
        "description": "更新商品请求"
      },
      "ApiResponseGoodsResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码：0=成功，1001=账号已存在，1002=账号不存在，1003=密码错误，1004=参数错误",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/GoodsResponse"
          }
        },
        "description": "通用响应"
      },
      "GoodsResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "商品ID",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "商品名称",
            "example": "测试商品"
          },
          "ctId": {
            "type": "integer",
            "description": "所属商户ID",
            "format": "int64",
            "example": 1
          }
        },
        "description": "商品响应"
      },
...
      "AddGoodsRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "商品名称",
            "example": "测试商品"
          }
        },
        "description": "添加商品请求"
      },
...
      "ApiResponseSessionResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码：0=成功，1001=账号已存在，1002=账号不存在，1003=密码错误，1004=参数错误",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/SessionResponse"
          }
        },
        "description": "通用响应"
      },
      "SessionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "会话ID",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "description": "用户ID",
            "format": "int64"
          },
          "ctId": {
            "type": "integer",
            "description": "商户ID",
            "format": "int64"
          },
          "goodsId": {
            "type": "integer",
            "description": "商品ID",
            "format": "int64"
          },
          "conversationStatus": {
            "type": "string",
            "description": "会话状态",
            "enum": [
              "ACTIVE",
              "CLOSED"
            ]
          },
          "lastMessageTime": {
            "type": "string",
            "description": "最后消息时间",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          },
          "lastMessageContent": {
            "type": "string",
            "description": "最后一条消息内容（预览）"
          },
          "unreadCount": {
            "type": "integer",
            "description": "当前用户在该会话中的未读消息数",
            "format": "int64"
          }
        },
        "description": "会话响应"
      },
      "ApiResponsePageMessageResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码：0=成功，1001=账号已存在，1002=账号不存在，1003=密码错误，1004=参数错误",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/PageMessageResponse"
          }
        },
        "description": "通用响应"
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "消息ID",
            "format": "int64"
          },
          "messageId": {
            "type": "string",
            "description": "消息唯一标识"
          },
          "sessionId": {
            "type": "integer",
            "description": "会话ID",
            "format": "int64"
          },
          "senderId": {
            "type": "integer",
            "description": "发送方ID",
            "format": "int64"
          },
          "senderType": {
            "type": "string",
            "description": "发送方类型",
            "enum": [
              "TENANT",
              "USER"
            ]
          },
          "content": {
            "type": "string",
            "description": "消息内容"
          },
          "isRead": {
            "type": "boolean",
            "description": "是否已读"
          },
          "readAt": {
            "type": "string",
            "description": "已读时间",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          }
        },
        "description": "消息响应"
      },
      "PageMessageResponse": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortObject"
            }
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        },
        "description": "响应数据"
      },
      "PageableObject": {
        "type": "object",
        "properties": {
          "paged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "unpaged": {
            "type": "boolean"
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortObject"
            }
          },
          "offset": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "direction": {
            "type": "string"
          },
          "nullHandling": {
            "type": "string"
          },
          "ascending": {
            "type": "boolean"
          },
          "property": {
            "type": "string"
          },
          "ignoreCase": {
            "type": "boolean"
          }
        }
      },
      "ApiResponsePageSessionResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码：0=成功，1001=账号已存在，1002=账号不存在，1003=密码错误，1004=参数错误",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/PageSessionResponse"
          }
        },
        "description": "通用响应"
      },
      "PageSessionResponse": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortObject"
            }
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        },
        "description": "响应数据"
      },
      "ApiResponsePageGoodsResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码：0=成功，1001=账号已存在，1002=账号不存在，1003=密码错误，1004=参数错误",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/PageGoodsResponse"
          }
        },
        "description": "通用响应"
      },
      "PageGoodsResponse": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortObject"
            }
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GoodsResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        },
        "description": "响应数据"
      },
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码：0=成功，1001=账号已存在，1002=账号不存在，1003=密码错误，1004=参数错误",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "success"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          }
        },
        "description": "通用响应"
      }
    }
  }
}